L12

+++ src/Car.java
@@ -3,7 +3,7 @@ public class Car {
     public String model = "";
     public int year = 0;
     public String vin = "";
-    public double price = 0.0;
+    public int price = 0;
     public String colour = "";
     public int mileage = 0;
     public String condition = "";
